Extract Wikidata HTTP helper and tighten read.ts types
Move fetchJsonWithRetries and shared search headers into wikidataHttp.ts.
Use safer entity/sitelink and claims typing in read.ts without changing
search behavior yet.
