+++ src/lib/wikidata/read.ts
@@ -1,8 +1,17 @@
-import { Entity, EntityId, ItemId, SearchResponse } from 'wikibase-sdk'
+import {
+  Entity,
+  EntityId,
+  ItemId,
+  SearchResponse,
+} from 'wikibase-sdk'
 import { Claim, transformFromWikidataDateStringToDate, wbk } from './util'
 import { WbGetEntitiesResponse } from 'wikibase-sdk/dist/src/helpers/parse_responses'
 import { SearchEntitiesOptions } from 'wikibase-sdk/dist/src/queries/search_entities'
 import wikidataConfig from '../../config/wikidata'
+import {
+  fetchJsonWithRetries,
+  WIKIDATA_SEARCH_HEADERS,
+} from './wikidataHttp'
 
 const {
   CARBON_FOOTPRINT,
@@ -14,57 +23,6 @@ const {
   ARCHIVE_URL,
 } = wikidataConfig.properties
 
-async function fetchJsonWithRetries<T = any>(
-  url: string,
-  {
-    headers,
-    maxAttempts = 3,
-    expectedContentType = 'application/json',
-    context,
-  }: {
-    headers?: Record<string, string>
-    maxAttempts?: number
-    expectedContentType?: string
-    context?: string
-  }
-): Promise<T> {
-  let attempt = 0
-  let res: Response | undefined
-  while (attempt < maxAttempts) {
-    res = await fetch(url, { headers })
-    if (res.ok) break
-    if ([429, 502, 503, 504].includes(res.status)) {
-      await new Promise((r) => setTimeout(r, (attempt + 1) * 1000))
-      attempt++
-      continue
-    }
-    break
-  }
-
-  const ctx = context ? `${context} ` : ''
-
-  if (!res) {
-    throw new Error(`${ctx}no response received`)
-  }
-
-  if (!res.ok) {
-    const text = await res.text().catch(() => '')
-    throw new Error(
-      `${ctx}HTTP ${res.status} ${res.statusText} – body: ${text.slice(0, 300)}`
-    )
-  }
-
-  const ct = (res.headers.get('content-type') || '').toLowerCase()
-  if (!ct.includes(expectedContentType)) {
-    const text = await res.text().catch(() => '')
-    throw new Error(
-      `${ctx}returned non-JSON (${ct}) – body: ${text.slice(0, 300)}`
-    )
-  }
-
-  return (await res.json()) as T
-}
-
 export async function getWikipediaTitle(id: EntityId): Promise<string> {
   const url = wbk.getEntities({
     ids: [id],
@@ -73,9 +31,9 @@ export async function getWikipediaTitle(id: EntityId): Promise<string> {
   const { entities }: WbGetEntitiesResponse = await fetch(url).then((res) =>
     res.json()
   )
-  const entity = entities[id] as any
-  const title =
-    entity?.sitelinks?.enwiki?.title ?? entity?.sitelinks?.svwiki?.title ?? null
+  const entity = entities[id]
+  const sitelinks = entity?.type === 'item' ? entity.sitelinks : undefined
+  const title = sitelinks?.enwiki?.title ?? sitelinks?.svwiki?.title ?? null
 
   if (!title) {
     throw new Error('No Wikipedia site link found')
@@ -99,7 +57,7 @@ export async function getLEINumber(
     return
   }
 
-  const claims = wikidataEntities[entity].claims
+  const { claims } = wikidataEntities[entity]
 
   if (
     claims === undefined ||
@@ -116,38 +74,32 @@ export async function searchCompany({
   companyName,
   language = 'sv',
 }: {
-  companyName
+  companyName: string
   language?: SearchEntitiesOptions['language']
 }): Promise<SearchResponse['search']> {
-  // TODO: try to search in multiple languages. Maybe we can find a page in English if it doesn't exist in Swedish?
   const searchEntitiesQuery = wbk.searchEntities({
     search: companyName,
     type: 'item',
-    // IDEA: Maybe determine language based on report or company origin. Or maybe search in multiple languages.
     language,
     limit: 20,
   })
 
-  const headers = {
-    Accept: 'application/json',
-    'User-Agent': 'KlimatkollenGarboBot/1.0 (+https://klimatkollen.se)',
-  }
-
-  const response = (await fetchJsonWithRetries<SearchResponse>(
+  const response = await fetchJsonWithRetries<SearchResponse>(
     searchEntitiesQuery,
     {
-      headers,
+      headers: { ...WIKIDATA_SEARCH_HEADERS },
       maxAttempts: 3,
       expectedContentType: 'application/json',
       context: 'Wikidata search',
     }
-  )) as SearchResponse
+  )
 
-  if ((response as any)?.error) {
-    throw new Error('Wikidata search failed: ' + (response as any).error)
+  if (response.error) {
+    const msg = response.error.info || response.error.code
+    throw new Error(`Wikidata search failed: ${msg}`)
   }
 
-  return response.search
+  return response.search ?? []
 }
 
 export async function getWikidataEntities(ids: `Q${number}`[]) {
@@ -155,14 +107,10 @@ export async function getWikidataEntities(ids: `Q${number}`[]) {
     ids,
     props: ['info', 'claims', 'descriptions', 'labels'],
   })
-  const headers = {
-    Accept: 'application/json',
-    'User-Agent': 'KlimatkollenGarboBot/1.0 (+https://klimatkollen.se)',
-  }
 
   const { entities }: WbGetEntitiesResponse =
     await fetchJsonWithRetries<WbGetEntitiesResponse>(url, {
-      headers,
+      headers: { ...WIKIDATA_SEARCH_HEADERS },
       maxAttempts: 3,
       expectedContentType: 'application/json',
       context: 'Wikidata entities',
@@ -192,7 +140,7 @@ export async function getClaims(entity: ItemId): Promise<Claim[]> {
       return []
     }
 
-    const claims = wikidataEntities[entity].claims
+    const { claims } = wikidataEntities[entity]
     if (claims === undefined) {
       return []
     }
@@ -205,30 +153,42 @@ export async function getClaims(entity: ItemId): Promise<Claim[]> {
 
       const getQualifierValue = (
         propertyId: string,
-        transformFn?: (value: any) => any
+        transformFn?: (value: unknown) => unknown
       ) => {
-        if (!claim.qualifiers || !claim.qualifiers[propertyId]) return ''
-        const value = claim.qualifiers[propertyId][0].datavalue.value
+        const snaks = claim.qualifiers?.[propertyId]
+        if (!snaks?.length) return ''
+        const [
+          {
+            datavalue: { value },
+          },
+        ] = snaks
         return transformFn ? transformFn(value) : value
       }
 
-      const getReferenceValue = (propertyId) => {
+      const getReferenceValue = (propertyId: string) => {
         if (!references || !references[propertyId]) return undefined
         return references[propertyId][0].datavalue.value
       }
 
       return {
         startDate: getQualifierValue(START_TIME, (value) =>
-          transformFromWikidataDateStringToDate(value.time)
+          transformFromWikidataDateStringToDate(
+            (value as { time: string }).time
+          )
         ),
         endDate: getQualifierValue(END_TIME, (value) =>
-          transformFromWikidataDateStringToDate(value.time)
+          transformFromWikidataDateStringToDate(
+            (value as { time: string }).time
+          )
         ),
         value: claim.mainsnak.datavalue.value.amount,
-        category: getQualifierValue(APPLIES_TO_PART, (value) => value.id),
+        category: getQualifierValue(
+          APPLIES_TO_PART,
+          (value) => (value as { id: ItemId }).id
+        ),
         scope: getQualifierValue(
           OBJECT_OF_STATEMENT_HAS_ROLE,
-          (value) => value.id
+          (value) => (value as { id: ItemId }).id
         ),
         id: claim.id,
         referenceUrl: getReferenceValue(REFERENCE_URL),
